Fix unset freeHGlobal delegate
This can happen because the static code of the NativeMethods class
could have not run, for some reason...
